pro-package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-of-agents"
version = "0.1.0"
description = "Persona management and workflow generation for product development"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Blake Aber" },
]
keywords = ["personas", "workflows", "product-development", "ux-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.39",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
voa = "voice_of_agents.cli.main:cli"

[project.entry-points."voice_of_agents.products"]

[tool.hatch.build.targets.wheel]
packages = ["src/voice_of_agents"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests requiring LLM",
]
